+++ src/tasks/merge-reports.ts
@@ -1,4 +1,60 @@
-import { EthGasReporterOutput } from './types';
+import fs from "fs"
+import path from "path"
+import { task, subtask } from "hardhat/config";
+import { EthGasReporterOutput } from '../types';
+import { TASK_GAS_REPORTER_MERGE, TASK_GAS_REPORTER_MERGE_REPORTS } from "../task-names";
+
+/**
+ * Task for merging multiple gasReporterOutput.json files generated by eth-gas-reporter
+ * This task is necessary when we want to generate different parts of the reports
+ * parallelized on different jobs, then merge the results and upload it to a CI service.
+ */
+
+subtask(TASK_GAS_REPORTER_MERGE_REPORTS)
+  .addOptionalVariadicPositionalParam(
+    "inputFiles",
+    "Path of several gasReporterOutput.json files to merge",
+    []
+  )
+  .setAction(async ({ inputFiles }: { inputFiles: string[] }) => {
+    const reports = inputFiles.map((input) => JSON.parse(fs.readFileSync(input, "utf-8")));
+    return mergeReports(reports);
+  })
+
+task(TASK_GAS_REPORTER_MERGE)
+  .addOptionalParam(
+    "output",
+    "Target file to save the merged report",
+    "gasReporterOutput.json"
+  )
+  .addVariadicPositionalParam(
+    "input",
+    "A list of gasReporterOutput.json files generated by eth-gas-reporter. Files can be defined using glob patterns"
+  )
+  .setAction(async (taskArguments, { run }) => {
+    const output = path.resolve(process.cwd(), taskArguments.output);
+
+    // Parse input files and calculate glob patterns
+    const { globSync } = await import("hardhat/internal/util/glob");
+    const arrayUniq = require("array-uniq");
+    const inputFiles = arrayUniq(taskArguments.input.map(globSync).flat())
+      .map(inputFile => path.resolve(inputFile));
+
+    if (inputFiles.length === 0) {
+      throw new Error(`No files found for the given input: ${taskArguments.input.join(" ")}`);
+    }
+
+    console.log(`Merging ${inputFiles.length} input files:`);
+    inputFiles.forEach(inputFile => {
+      console.log("  - ", inputFile);
+    });
+
+    console.log("\nOutput: ", output);
+
+    const result = await run(TASK_GAS_REPORTER_MERGE_REPORTS, { inputFiles });
+
+    fs.writeFileSync(output, JSON.stringify(result), "utf-8");
+  });
 
 /**
  * Try to merge several gas reporter output objects into one. It will also
